KSH 카트 레저 수량 수정 커밋

+++ project_sample/sample1/src/main/java/com/example/sample1/model/Admin.java
@@ -47,7 +47,7 @@ public class Admin {
 	private String leisureName;
 	private int leisureNo;
 	private int leisurePrice;
-	private int leisureSales;
+	private double leisureSales;
 	
 	//inner
 	private String uId;
